# Optimized constant `SignalProducer`s.