These speedups are applicable to only closure-based observation APIs (i.e. `startWithSignal` is ineligible).

| Item | testSequence (32 items) | testValue |
| --- | --- | --- |
| `master` | avg 30296 ns; min 26426 ns | avg 8910 ns; min 8171 ns |
| `unsafeAction` | avg 16092 ns; min 14627 ns | avg 2928 ns; min 2568 ns |
| +`unsafeUninterruptibleAction` | avg 16140 ns; min 14502 ns | avg 664 ns; min 587 ns |
| Speedup | ~1.85x | ~13.5x |

[Benchmark source](https://gist.github.com/andersio/8d519b366aace3e85063ef1a2cbb38b1).

This PR also relaxed `UnsafeAtomicState.is`.

#### Checklist
- [ ] Updated CHANGELOG.md.