@andersio 2017-07-16
Probably need a larger-scale change for this to be picked up by everywhere, most importantly the flattening operators.