@andersio 2017-07-16
Some interesting numbers:
```
@testFlatMapSequence(): avg 152921 ns; min 137964 ns
@testSequence(): avg 15598 ns; min 14388 ns

@testValue(): avg 660 ns; min 591 ns
@testFlatMapValue(): avg 54609 ns; min 49320 ns
```

:/